refactor(console): extract LANGUAGE_LIST to a side-effect-free module
Address PR review feedback: importing LANGUAGE_LIST from the LanguageSwitcher
component module pulled its UI-only dependencies (Dropdown, index.module.less)
into every consumer and coupled the sidebar layout to a specific component
just to read a constant.

Move LANGUAGE_LIST (and the LanguageConfig type) into a dedicated,
side-effect-free module, constants/languageList.tsx, which imports only icons.
SidebarSettingsPanel and LanguageSwitcher now import the list from there.
LanguageSwitcher re-exports it for backward compatibility, so Header - which
already renders the LanguageSwitcher component - keeps working unchanged.
